app/requirements.txt
streamlit
pandas
langchain
langchain-google-genai
rapidfuzz
duckdb
sqlalchemy
scikit-learn
python-dotenv
